Required changes for test and train.
- Train should not cause any breaking changes.
- Test may cause breaking changes - it assumes that you need to provide
  a dictionary of lists, where each key represents a completely new
model (vdn, or independent_ql, etc) and the list represents each of the
variations of that model if there where multiple trained from a single
test config. A simple function can be created which could start as an
entrypoint for experiments when not using the backend, which would
formulate this dictionary of lists using the previous hardcoded naming
convention.